Update error return values in Load and GetSheet
